group: unbounded key admission through the v2 hash path
agg_v2_can_handle drops its 16-key cap; the five [16] buffers it
protected become carves (the radix scatter packs keys straight into
the record — no per-row buffer). by-dict and sym-vector distinct
gates lift with dep_key_names carved. Legacy group.c stays 8-gated
(cut 4); dense planners keep 1..16 as routing.

+++ src/ops/agg_engine.h
@@ -29,17 +29,10 @@ typedef struct {
  * order == first-occurrence order; first_row[gid] records the row where the
  * group first appeared. Returns 0 on success (caller releases out via
  * agg_groups_free()), -1 on allocation failure.
- * n_keys stays uint8_t: the interface admits up to 255, but the LIVE bound
- * today is 16 via two independent admission gates — the GROUP path via
- * agg_v2_can_handle's 1..16-key admission, and the keys-only DISTINCT path
- * via query.c's own `nk > 16` distinct-admission gate (~query.c:6338).
- * NOTE: the by-dict's "1..16 keys" cap in query.c (~5203) does NOT cover
- * this — it only guards RAY_DICT by: shapes; a raw sym-vector by: bypasses
- * it and previously overran the fixed data[16] inside agg_group_keys at 17
- * keys (fixed by adding the dedicated gate above it). Cut-3 is the boundary
- * where either gate may lift past 16, at which point the fixed data[16]
- * inside agg_group_keys must become a carve. */
-int agg_group_keys(ray_t** key_cols, uint8_t n_keys, int64_t nrows, agg_groups_t* out);
+ * n_keys is uint32_t: unbounded key count.  Cut-3 lifted both admission gates
+ * (the GROUP path and the keys-only DISTINCT path) and the fixed data[16]
+ * inside became an exact carve, so any key count groups correctly. */
+int agg_group_keys(ray_t** key_cols, uint32_t n_keys, int64_t nrows, agg_groups_t* out);
 
 /* Release the buffers an agg_groups_t holds (buddy-backed, NOT libc malloc — so
  * callers must use this, not free()).  Idempotent; NULLs the pointers. */
@@ -52,12 +45,10 @@ void agg_groups_free(agg_groups_t* out);
  * and every tbl column is fixed-width/SYM/STR/LIST.  Caller owns the table.
  * keep_syms (or NULL) lists the column syms a consumer references — non-key
  * columns NOT in it are dropped (projection pushdown); keys are always kept.
- * nk stays uint8_t: query.c's only caller gates `nk > 16` before the call
- * (~query.c:6338 — falls through to the composite path beyond 16 keys, the
- * live bound of agg_group_keys' fixed data[16]), so this parameter never
- * sees a truncated value. */
+ * nk is uint32_t: unbounded key count (cut-3 lifted query.c's distinct gate;
+ * agg_group_keys carves its key table, so any nk groups correctly). */
 ray_t* agg_select_distinct(ray_t* tbl, ray_t** key_cols, const int64_t* key_syms,
-                           uint8_t nk, int64_t nrows,
+                           uint32_t nk, int64_t nrows,
                            const int64_t* keep_syms, int keep_n);
 
 /* Build a dense SoA per-group state array for one aggregate (vt), run a single
@@ -76,10 +67,10 @@ ray_t* agg_run_one(const agg_vtable_t* vt, ray_t* val_col,
 
 typedef struct {
     bool     ok;
-    uint32_t n_keys;        /* mirrors ext->n_keys' width; value stays 1..16 */
-    int64_t  mins[16];      /* [16]: protected by agg_v2_can_handle's 1..16-key admission (cut-3 boundary) */
-    int64_t  ranges[16];    /* [16]: protected by agg_v2_can_handle's 1..16-key admission (cut-3 boundary) */
-    int64_t  strides[16];   /* [16]: protected by agg_v2_can_handle's 1..16-key admission (cut-3 boundary); composite packing: slot = sum_k (key_k - min_k)*strides[k] */
+    uint32_t n_keys;        /* mirrors ext->n_keys' width; value stays 1..16 (dense self-limit) */
+    int64_t  mins[16];      /* [16]: dense direct-index routing self-limits to <=16 keys (agg_dense_plan) */
+    int64_t  ranges[16];    /* [16]: dense direct-index routing self-limits to <=16 keys (agg_dense_plan) */
+    int64_t  strides[16];   /* [16]: dense self-limit <=16; composite packing: slot = sum_k (key_k - min_k)*strides[k] */
     int64_t  total_slots;   /* product of ranges */
 } dense_plan_t;
 
@@ -88,11 +79,12 @@ typedef struct {
  *  - every agg vtable is ACC_STREAMING (no buffered median/top)
  *  - product of per-key ranges <= DENSE_MAX_SLOTS (no overflow)
  * Does one min/max prescan over the key columns.  Sets out->ok accordingly.
- * n_keys stays uint8_t: only ever called with ext->n_keys post agg_v2_can_handle's
- * 1..16 admission (dense direct-index routing bound; wider shapes take the hash
- * path) — see agg_v2_can_handle's comment.  n_aggs is unused by dense eligibility
- * (kept uint32_t so a widened ext->n_aggs never narrows at this call). */
-bool agg_dense_plan(ray_t** key_cols, uint8_t n_keys,
+ * n_keys is uint32_t (untruncated ext->n_keys): dense direct-index routing
+ * self-limits to <=16 keys here (wider shapes are rejected to v2's unbounded
+ * hash/radix path), so the fixed [16] mins/ranges/strides are only ever read
+ * at <=16.  n_aggs is unused by dense eligibility (kept uint32_t so a widened
+ * ext->n_aggs never narrows at this call). */
+bool agg_dense_plan(ray_t** key_cols, uint32_t n_keys,
                     const agg_vtable_t** vts, uint32_t n_aggs,
                     int64_t nrows, dense_plan_t* out);
